Minimize circular dependencies
