+++ src/utils/terms.ts
@@ -29,15 +29,6 @@ export function isNamedNode(obj): obj is NamedNode {
   return isTFTerm(obj) && obj.termType === 'NamedNode'
 }
 
-/** Retrieve the value of a term, or self if already a string. */
-export function termValue(node: TFTerm | string): string {
-  if (typeof node === 'string') {
-    return node
-  }
-
-  return node.value
-}
-
 /** TypeGuard for RDFLib Collections */
 export function isCollection(obj: any): obj is Collection<any> {
   return isTFTerm(obj)
