+++ tests/unit/parse-test.js
@@ -3,8 +3,8 @@ import { expect } from 'chai'
 
 import parse from '../../src/parse'
 import CanonicalDataFactory from '../../src/data-factory-internal'
-import DataFactory from '../../src/data-factory'
 import defaultXSD from '../../src/xsd'
+import DataFactory from '../../src/rdflib-data-factory'
 
 describe('Parse', () => {
   describe('ttl', () => {
